requirements.txt
chromadb==1.4.1
llama-index-core==0.14.13
llama-index-cli==0.5.3
llama-index-embeddings-ollama==0.8.6
llama-index-llms-ollama==0.9.1
llama-index-vector-stores-chroma==0.5.5
llama-index-readers-file==0.5.6
llama-index-workflows==2.13.0
ollama==0.6.1
pandas==2.3.3
